Rename lib geolocation to country-code
