+++ src/lib/header-bidding/utils.spec.ts
@@ -1,8 +1,8 @@
 import type { CountryCode } from '@guardian/libs';
 import { createAdSize } from '../../core/ad-sizes';
 import { isUserInVariant as isUserInVariant_ } from '../../experiments/ab';
+import { getCountryCode as getCountryCode_ } from '../../utils/country-code';
 import { _ } from '../../utils/geo-utils';
-import { getCountryCode as getCountryCode_ } from '../../utils/geolocation';
 import type { SourceBreakpoint } from '../detect/detect-breakpoint';
 import {
 	getCurrentTweakpoint as getCurrentTweakpoint_,
