+++ src/utils/geo-utils.ts
@@ -1,5 +1,5 @@
 import type { CountryCode } from '@guardian/libs';
-import { getCountryCode } from './geolocation';
+import { getCountryCode } from './country-code';
 
 // cache the users location so we only have to look it up once
 let geo: CountryCode | undefined;
